--- packages/rest-typings/src/v1/e2e/e2eGetUsersOfRoomWithoutKeyParamsGET.ts
@@ -1,20 +1,0 @@
-import { ajv } from '../Ajv';
-
-export type e2eGetUsersOfRoomWithoutKeyParamsGET = {
-	rid: string;
-};
-
-const e2eGetUsersOfRoomWithoutKeyParamsGETSchema = {
-	type: 'object',
-	properties: {
-		rid: {
-			type: 'string',
-		},
-	},
-	additionalProperties: false,
-	required: ['rid'],
-};
-
-export const ise2eGetUsersOfRoomWithoutKeyParamsGET = ajv.compile<e2eGetUsersOfRoomWithoutKeyParamsGET>(
-	e2eGetUsersOfRoomWithoutKeyParamsGETSchema,
-);
